% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignTaxonomyCustomMock.R
\name{assignTaxonomyCustomMock}
\alias{assignTaxonomyCustomMock}
\title{Assign Custom Mocks Taxonomy}
\usage{
assignTaxonomyCustomMock(
  x,
  mock_db = NULL,
  processors = NULL,
  threshold = 60,
  strand = "top",
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Phyloseq objects where taxa_names are ASV sequences}

\item{mock_db}{ZymoTrainingSet}

\item{processors}{Passed on to DECIPHER::IdTaxa Default is NULL}

\item{threshold}{Passed on to DECIPHER::IdTaxa. Default is 60}

\item{strand}{DECIPHER::IdTaxa Default is top}

\item{verbose}{Default is FALSE}

\item{...}{Arguments to pass on to DECIPHER::IdTaxa}
}
\value{
Phyloseq with new taxonomy
}
\description{
Assign Custom Mocks Taxonomy
}
\details{
For consistency in taxonomic assignments, use only a small database
         that has 16S rRNA gene sequences from mock community genomes.
         This is done to avoid any clashes/changes that happen with public databases
         and to get a direct comparison.
}
\examples{
#output.dat <- assignTaxonomyCustomMock(ps.zym,
 #                                      mock_db = NULL,
  #                                     processors = 2,
   #                                    threshold = 80)
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
