% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkZymoBiomics.R
\name{checkZymoBiomics}
\alias{checkZymoBiomics}
\title{Compare ZymoBiomics to Experimental}
\usage{
checkZymoBiomics(
  x,
  mock_db = NULL,
  multithread = 2,
  threshold = 60,
  verbose = FALSE,
  strand = "top",
  ...
)
}
\arguments{
\item{x}{Phyloseq objects where taxa_names are ASV sequences}

\item{mock_db}{ZymoTrainingSet}

\item{multithread}{Passed on to DECIPHER::IdTaxa Default is NULL}

\item{threshold}{Passed on to DECIPHER::IdTaxa. Default is 60}

\item{verbose}{Default is FALSE}

\item{strand}{DECIPHER::IdTaxa Default is top}

\item{...}{Arguments to pass on to DECIPHER::IdTaxa}
}
\value{
List with pseq, and cor table
}
\description{
Compare ZymoBiomics to Experimental
}
\details{
Compare ZymoBiomics theoretical to Experimental composition.
}
\examples{
library(phyloseq)

output.dat <- checkZymoBiomics(ZymoExamplePseq,
                               mock_db = ZymoTrainingSet,
                               multithread= 2,
                               threshold = 80,
                               strand = "top",
                               verbose = FALSE)

}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
