% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare2theorectical.R
\name{compare2theorectical}
\alias{compare2theorectical}
\title{Compare Custom Mocks to Experimental}
\usage{
compare2theorectical(x, theoretical_id)
}
\arguments{
\item{x}{Phyloseq object with experimental and theoretical composition}

\item{theoretical_id}{name of the theoretical sample. Must be in sample_names(x)}
}
\value{
Correlation table
}
\description{
Compare Custom Mocks to Experimental
}
\details{
Correlation between theoretical and experimental mocks.
}
\examples{
library(phyloseq)
ps.zym <- ZymoExamplePseq
taxa_names(ps.zym) <- refseq(ps.zym)
compare2theorectical(ps.zym, theoretical_id = "somename")

}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
