% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotZymoDefault.R
\name{plotZymoDefault}
\alias{plotZymoDefault}
\title{Plot ZymoBiomics to Experimental comparison}
\usage{
plotZymoDefault(x)
}
\arguments{
\item{x}{Output from \code{checkZymoBiomics}}
}
\value{
ggplot object
}
\description{
Plot ZymoBiomics to Experimental comparison
}
\details{
A wrapper to plot the output of \code{checkZymoBiomics}.
}
\examples{
library(phyloseq)
output.dat <- checkZymoBiomics(ZymoExamplePseq,
                               mock_db = ZymoTrainingSet,
                               multithread= 2,
                               threshold = 80,
                               strand = "top",
                               verbose = FALSE)
plotZymoDefault(output.dat)

}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
