% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZymoBiomics}
\alias{ZymoBiomics}
\title{ZymoBiomics}
\format{
Phyloseq object
}
\usage{
data(ZymoBiomics)
}
\value{
Loads the data set in R.
}
\description{
ZymoBiomics data from ZymoBIOMICS Microbial Community Standard.
}
\details{
ZymoBiomics data from ZymoBIOMICS Microbial Community Standard.
         Has a table with  information on Species and their theoretical
         composition in percentage values for Genomic DNA, 16S Only,
         16S and 18S, Genome Copy and Cell number. Product name D6300
}
\references{
\itemize{
\item{}{ZymoBIOMICS™ Microbial Community Standard.
       \emph{Catalog No. D6300.
       \url{https://files.zymoresearch.com/protocols/_d6300_zymobiomics_microbial_community_standard.pdf}
}
}
}
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
