% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZymoBiomicsFasta}
\alias{ZymoBiomicsFasta}
\title{ZymoBiomicsFasta}
\format{
DNAStringSet object
}
\usage{
data(ZymoBiomicsFasta)
}
\value{
Loads the data set in R.
}
\description{
16S and 18S sequences from ZymoBIOMICS Microbial Community Standard.
}
\details{
ZymoBiomicsFasta has all 16S and 18S sequence data as \code{DNAStringSet}.
         Has 50 rDNA sequences, all copies of 16S rRNA gene are made available.
}
\references{
\itemize{
\item{}{ZymoBIOMICS™ Microbial Community Standard.
       \emph{Catalog No. D6300.
       \url{https://files.zymoresearch.com/protocols/_d6300_zymobiomics_microbial_community_standard.pdf}
       \url{https://s3.amazonaws.com/zymo-files/BioPool/ZymoBIOMICS.STD.refseq.v2.zip}
}
}
}
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
