% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZymoBiomicsGut}
\alias{ZymoBiomicsGut}
\title{ZymoBiomicsGut}
\format{
Table
}
\usage{
data(ZymoBiomicsGut)
}
\value{
Loads the data set in R.
}
\description{
Contains ZymoBiomicsGut compatible with DECIPHER::IdTaxa.
}
\details{
ZymoBiomicsGut contains table with information on strains and their
         abundances ZymoBIOMICS Gut Microbial Community Standard Catalog
         No. D6331.
}
\references{
\itemize{
\item{}{ZymoBIOMICS® Gut Microbiome Standard.
       \emph{Catalog No. D6331.
       \url{https://files.zymoresearch.com/protocols/_d6331_zymobiomics_gut_microbiome_standard.pdf}
}
}
}
Shetty SA 2021
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
