% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZymoBiomicsGutPseq}
\alias{ZymoBiomicsGutPseq}
\title{ZymoBiomicsGutPseq}
\format{
Phyloseq object
}
\usage{
data(ZymoBiomicsGutPseq)
}
\value{
Loads the data set in R.
}
\description{
Theoretical 16S composition of ZymoBIOMICS Microbial Community Standard.
}
\details{
ZymoBiomicsGutPseq contains the \emph{theoretical} 16S composition of
         ZymoBIOMICS Gut Microbial Community Standard. One sample, 19 known taxa and 7
         taxonomic ranks.
}
\references{
\itemize{
\item{}{ZymoBIOMICS® Gut Microbiome Standard.
       \emph{Catalog No. D6331.
       \url{https://files.zymoresearch.com/protocols/_d6331_zymobiomics_gut_microbiome_standard.pdf}
}
}
}
Shetty SA 2021
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
