% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZymoBiomicsGutTrainingSet}
\alias{ZymoBiomicsGutTrainingSet}
\title{ZymoBiomicsGutTrainingSet}
\format{
TrainingSet object
}
\usage{
data(ZymoBiomicsGutTrainingSet)
}
\value{
Loads the data set in R.
}
\description{
Contains ZymoBiomicsGutTrainingSet compatible with DECIPHER::IdTaxa.
}
\details{
ZymoBiomicsGutTrainingSet contains 16S sequences of
         ZymoBIOMICS Gut Microbial Community Standard Catalog No. D6331.
}
\references{
\itemize{
\item{}{ZymoBIOMICS® Gut Microbiome Standard.
       \emph{Catalog No. D6331.
       \url{https://files.zymoresearch.com/protocols/_d6331_zymobiomics_gut_microbiome_standard.pdf}
}
}
}
Shetty SA 2021
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
