% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZymoBiomicsPseq}
\alias{ZymoBiomicsPseq}
\title{ZymoBiomicsPseq}
\format{
Phyloseq object
}
\usage{
data(ZymoBiomicsPseq)
}
\value{
Loads the data set in R.
}
\description{
Theoretical 16S composition of ZymoBIOMICS Microbial Community Standard.
}
\details{
ZymoBiomicsPseq contains the \emph{theoretical} 16S composition of
         ZymoBIOMICS Microbial Community Standard. One sample, 8 taxa and 7
         taxonomic ranks.
}
\references{
\itemize{
\item{}{ZymoBIOMICS™ Microbial Community Standard.
       \emph{Catalog No. D6300.
       \url{https://files.zymoresearch.com/protocols/_d6300_zymobiomics_microbial_community_standard.pdf}
}
}
}
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
