% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZymoExamplePseq}
\alias{ZymoExamplePseq}
\title{ZymoExamplePseq}
\format{
Phyloseq object
}
\usage{
data(ZymoExamplePseq)
}
\value{
Loads the data set in R.
}
\description{
Contains ZymoBIOMICS Microbial Community Standard samples ran
             in different sequencing libraries as postive controls.
}
\details{
ZymoExamplePseq contains the \emph{experimental} 16S composition of
         ZymoBIOMICS Microbial Community Standard. 9 samples, 946 ASVs and 7
         taxonomic ranks. ZymoBIOMIC mock community standards consisting of
         eight bacterial species \emph{Pseudomonas.aeruginosa},
         \emph{Escherichia coli}, \emph{Salmonella.enterica},
         \emph{Lactobacillus.fermentum}, \emph{Enterococcus.faecalis},
         \emph{Staphylococcus.aureus}, \emph{Listeria.monocytogenes},
         \emph{Bacillus.subtilis} and two fungal species
         \emph{Saccharomyces.cerevisiae} and \emph{Cryptococcus.neoformans}.
         The dilutions were done in microbial free water (Qiagen) in
         eight rounds of a serial three-fold dilution
         prior to DNA extraction. The data are for the V4 region of the
         16S rRNA gene profiled with Illumina MiSeq.
}
\references{
\itemize{
\item{}{Karstens L, Asquith M, Davin S, Fair D, Gregory WT, Wolfe AJ,
        Braun J, McWeeney S. (2019) Controlling for contaminants in
        low-biomass 16S rRNA gene sequencing experiments.
       \emph{mSystems}, 4:e00290-19.
       \url{https://doi.org/10.1128/mSystems.00290-19}
}
}
Shetty SA 2021
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
