% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZymoTrainingSet}
\alias{ZymoTrainingSet}
\title{ZymoTrainingSet}
\format{
TrainingSet object
}
\usage{
data(ZymoTrainingSet)
}
\value{
Loads the data set in R.
}
\description{
Contains ZymoTrainingSet compatible with DECIPHER::IdTaxa.
}
\details{
ZymoTrainingSet contains 16S sequences of
         ZymoBIOMICS Microbial Community Standard. ZymoBIOMIC mock community
         standards consisting of
         eight bacterial species \emph{Pseudomonas.aeruginosa},
         \emph{Escherichia coli}, \emph{Salmonella.enterica},
         \emph{Lactobacillus.fermentum}, \emph{Enterococcus.faecalis},
         \emph{Staphylococcus.aureus}, \emph{Listeria.monocytogenes},
         \emph{Bacillus.subtilis} and two fungal species
         \emph{Saccharomyces.cerevisiae} and \emph{Cryptococcus.neoformans}.
}
\references{
\itemize{
\item{}{ZymoBIOMICS™ Microbial Community Standard.
       \emph{Catalog No. D6300.
       \url{https://files.zymoresearch.com/protocols/_d6300_zymobiomics_microbial_community_standard.pdf}
}
}
}
Shetty SA 2021
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
