% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignTaxonomyCustomMock.R
\name{assignTaxonomyCustomMock}
\alias{assignTaxonomyCustomMock}
\title{Assign Custom Mocks Taxonomy}
\usage{
assignTaxonomyCustomMock(
  x,
  mock_db = NULL,
  multithread = multithread,
  minBoot = multithread,
  ...
)
}
\arguments{
\item{x}{A phyloseq object with experimental and theoretical composition}

\item{mock_db}{Path to custom database}

\item{multithread}{Passed on to dada2::assignTaxonomy Default is 2}

\item{minBoot}{Passed on to dada2::assignTaxonomy. Default is 80}

\item{...}{Arguments to pass on to dada2::assignTaxonomy}
}
\value{
Correlation table
}
\description{
Assign Custom Mocks Taxonomy
}
\details{
For consistency in taxonomic assignments, use only a small database
         that has 16S rRNA gene sequences from mock community genomes.
         This is done to avoid any clashes/changes that happen with public databases
         and to get a direct comparison.
}
\examples{
#output.dat <- assignTaxonomyCustomMock(ps.zym,
 #                                      mock_db = NULL,
  #                                     multithread= 2,
   #                                    minBoot = 80)
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
