% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkZymoBiomics.R
\name{checkZymoBiomics}
\alias{checkZymoBiomics}
\title{Compare ZymoBiomics to Experimental}
\usage{
checkZymoBiomics(x, mock_db = NULL, multithread = 2, minBoot = 80, ...)
}
\arguments{
\item{x}{Phyloseq objects where taxa_names are ASV sequences}

\item{mock_db}{Default is NULL and uses ZymoBiomics sequences}

\item{multithread}{Passed on to dada2::assignTaxonomy Default is 2}

\item{minBoot}{Passed on to dada2::assignTaxonomy. Default is 80}

\item{...}{Arguments to pass on to dada2::assignTaxonomy}
}
\value{
List with pseq, plots and cor table
}
\description{
Compare ZymoBiomics to Experimental
}
\details{
A wrapper to build phylogenetic tree for ASV sequences
         obtained after \code{dada2}.
}
\examples{
library(phyloseq)
ps.zym <- ZymoExamplePseq
taxa_names(ps.zym) <- refseq(ps.zym)
output.dat <- checkZymoBiomics(ps.zym,
                               mock_db = NULL,
                               multithread= 2,
                               minBoot = 80)
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
