% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZymoExamplePseq}
\alias{ZymoExamplePseq}
\title{ZymoExamplePseq}
\format{
Phyloseq object
}
\usage{
data(ZymoExamplePseq)
}
\value{
Loads the data set in R.
}
\description{
Contains ZymoBIOMICS Microbial Community Standard samples ran
             in different sequencing libraries as postive controls.
}
\details{
ZymoExamplePseq contains the \emph{experimental} 16S composition of
         ZymoBIOMICS Microbial Community Standard. 19 sample, 390 ASVs and 7
         taxonomic ranks.
}
\references{
Shetty SA 2021
}
\author{
Sudarshan Shetty \email{sudarshanshetty9@gmail.com}
}
\keyword{Data}
